requirements.txt
fastapi
stravalib
authlib
python-dotenv
uvicorn
httpx
itsdangerous